pyproject.toml
[project]
name = "devlens"
version = "0.1.0"
description = "Local-first developer skill analysis tool"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "libcst>=1.8.5",
    "ollama>=0.6.0",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.11.3",
    "radon>=6.0.1",
    "sqlalchemy>=2.0.43",
    "structlog>=25.4.0",
    "typer>=0.19.2",
]

[project.scripts]
devlens = "devlens.main:main"

[dependency-groups]
dev = [
    "mypy>=1.20.0",
    "pytest>=9.0.3",
    "ruff>=0.15.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/devlens"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
packages = ["devlens"]

[[tool.mypy.overrides]]
module = ["radon.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["qdrant_client"]
ignore_missing_imports = true
